requirements.txt
flask
pypdf
python-docx
Pillow
img2pdf
pdf2image
